segrega las validaciones de productos
